Cleanup/fix record trace menu/dialog
